+++ modules/common/radio.js
@@ -1,11 +1,20 @@
 
 
-// Whether or not we're in debug mode or not. debug mode helps you
-// get around the issues of lack of warnings when events are mis-typed.
-let shouldDebug = false;
+function createDebug() {
+  // Debug mode warns about overwritten or unhandled requests.
+  let shouldDebug = false;
 
-function setDebug(setShouldDebug = true) {
-  shouldDebug = setShouldDebug;
+  function setDebug(setShouldDebug = true) {
+    shouldDebug = setShouldDebug;
+  }
+
+  function debugLog(warning, eventName, channelName) {
+    if (shouldDebug && console && console.warn) {
+      console.warn(debugText(warning, eventName, channelName));
+    }
+  }
+
+  return { debugLog, setDebug };
 }
 
 // Format debug text.
@@ -14,15 +23,7 @@ function debugText(warning, eventName, channelName) {
     `: "${ eventName }"`;
 }
 
-// This is the method that's called when an unregistered event was called.
-// By default, it logs warning to the console. By overriding this you could
-// make it throw an Error, for instance. This would make firing a nonexistent event
-// have the same consequence as firing a nonexistent method on an Object.
-function debugLog(warning, eventName, channelName) {
-  if (shouldDebug && console && console.warn) {
-    console.warn(debugText(warning, eventName, channelName));
-  }
-}
+const { debugLog, setDebug } = createDebug();
 
 // Log information about the channel and event
 function log(channelName, eventName, ...args) {
@@ -32,6 +33,7 @@ function log(channelName, eventName, ...args) {
 }
 
 export {
+  createDebug,
   setDebug,
   debugLog,
   log,